requirements.txt
colorama==0.4.4
commonmark==0.9.1
einops==0.4.1
joblib==1.1.0
numpy==1.21.5
Pillow==7.1.2
Pygments==2.11.2
pymeshlab==2022.2
pytorch-metric-learning==1.2.0
rich==10.16.1
scikit-learn==1.0.1
scipy==1.7.3
threadpoolctl==3.1.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
typing_extensions==4.1.1
